# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce Store API",
    "description": "This API handles operations for an e-commerce store, including cart management, checkout processes, and admin functionalities.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/checkout": {
      "post": {
        "summary": "Checkout the cart and place an order",
        "operationId": "checkout",
        "tags": ["Checkout"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cartItems": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/CartItem"
                    }
                  },
                  "discountCode": {
                    "type": "string"
                  }
                },
                "required": ["cartItems"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderConfirmation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CartItem": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "example": 2
          }
        },
        "required": ["itemId", "quantity"]
      },
...
      "OrderConfirmation": {
        "type":        "object",
        "properties": {
          "orderId": {
            "type": "string",
            "example": "ORD123456"
          },
          "totalPrice": {
            "type": "number",
            "format": "float",
            "example": 100.50
          },
          "discountApplied": {
            "type": "boolean",
            "example": true
          },
          "finalPrice": {
            "type": "number",
            "format": "float",
            "example": 90.45
          }
        }
      },
...
    }
  }
}